requirements.txt
python-binance==1.0.19
streamlit==1.28.0
python-dotenv==1.0.0